requirements.txt
transformers
torch
streamlit
python-dotenv
speechrecognition
PyPDF2
python-docx
sentencepiece
protobuf
accelerate
